+++ internal/workerutil/main_test.go
@@ -1,0 +1,22 @@
+package workerutil
+
+import (
+	"flag"
+	"os"
+	"testing"
+
+	"github.com/inconshreveable/log15"
+	"github.com/sourcegraph/sourcegraph/internal/db/dbtesting"
+)
+
+func init() {
+	dbtesting.DBNameSuffix = "workerutil"
+}
+
+func TestMain(m *testing.M) {
+	flag.Parse()
+	if !testing.Verbose() {
+		log15.Root().SetHandler(log15.DiscardHandler())
+	}
+	os.Exit(m.Run())
+}
